requirements.txt
wheel
diff-match-patch
ipython
jinja2
matplotlib
natsort
pandas
pathlib
progressbar2
pyelftools
pyyaml